24786: handle read -t properly in zle via getbyte()

+++ Src/Zle/zle_keymap.c
@@ -1363,7 +1363,7 @@ getkeymapcmd(Keymap km, Thingy *funcp, char **strp)
 static int
 getkeybuf(int w)
 {
-    int c = getbyte(w, NULL);
+    int c = getbyte((long)w, NULL);
 
     if(c < 0)
 	return EOF;
